+++ controllers/genreController.js
@@ -28,10 +28,23 @@ const createGenre = async (req, res) => {
   res.redirect("/genre");
 };
 
-const deleteGenre = async (req, res) => {
+const renderDeleteGenre = async (req, res) => {
   const id = req.params.id;
 
-  await db.deleteGenre(id);
+  res.render(`delete/genreDelete`, {
+    id: id,
+  });
+};
+
+const deleteGenre = async (req, res) => {
+  const { password } = req.body;
+  const id = req.params.id;
+
+  if (password === process.env.ADMIN_PASSWORD) {
+    await db.deleteGenre(id);
+  } else {
+    console.log("Nope");
+  }
 
   res.redirect("/genre");
 };
@@ -42,4 +55,5 @@ module.exports = {
   renderGenreForm,
   createGenre,
   deleteGenre,
+  renderDeleteGenre,
 };
